+++ App.js
@@ -9,49 +9,22 @@ import {
   StatusBar,
   TouchableOpacity
 } from "react-native";
-// import { Font } from "expo";
 import * as Font from "expo-font";
+import { createAppContainer } from "react-navigation";
+import { createStackNavigator } from "react-navigation-stack";
+
+import Home from "./screens/Home";
+import AllGiants from "./screens/AllGiants";
+import { withNavigation } from "react-navigation";
+
+const RootStack = createStackNavigator({
+  Home: { screen: Home },
+  AllGiants: { screen: AllGiants }
+});
 
 export default class App extends React.Component {
-  state = {
-    assetsLoaded: false
-  };
-
-  async componentDidMount() {
-    await Font.loadAsync({
-      "amatic-sc": require("./assets/fonts/amatic-sc.ttf"),
-      "Satisfy-Regular": require("./assets/fonts/Satisfy-Regular.ttf")
-    });
-    this.setState({ assetsLoaded: true });
-  }
-
   render() {
-    const { assetsLoaded } = this.state;
-
-    if (assetsLoaded) {
-      return (
-        <ImageBackground
-          source={require("./bg.png")}
-          style={{ width: "100%", height: "100%" }}
-        >
-          <View style={styles.container}>
-            <Text style={{ fontFamily: "amatic-sc", fontSize: 60 }}>
-              The Forgotten Giants
-            </Text>
-            <TouchableOpacity style={styles.customBtnBG} onPress={() => {}}>
-              <Text style={styles.customBtnText}>Begin!</Text>
-            </TouchableOpacity>
-          </View>
-        </ImageBackground>
-      );
-    } else {
-      return (
-        <View>
-          <ActivityIndicator size="large" color="#0000ff" />
-          <StatusBar barStyle="default" />
-        </View>
-      );
-    }
+    return <Home />;
   }
 }
 
@@ -62,27 +35,5 @@ const styles = StyleSheet.create({
     flex: 1,
     alignItems: "center",
     justifyContent: "center"
-  },
-  buttonContainer: {
-    fontFamily: "Satisfy-Regular",
-    backgroundColor: "#D48104",
-    borderRadius: 5,
-    padding: 5
-  },
-  /* Here style the text of your button */
-  customBtnText: {
-    fontSize: 35,
-    fontWeight: "200",
-    color: "white",
-    fontFamily: "Satisfy-Regular"
-  },
-
-  /* Here style the background of your button */
-  customBtnBG: {
-    marginTop: 30,
-    backgroundColor: "#D48104",
-    paddingHorizontal: 10,
-    paddingVertical: 5,
-    borderRadius: 5
   }
 });
