+++ screens/AllGiants.js
@@ -1,0 +1,11 @@
+import React from "react";
+import { View, Text } from "react-native";
+import { createAppContainer } from "react-navigation";
+import { createStackNavigator } from "react-navigation-stack";
+import { withNavigation } from "react-navigation";
+
+export default class AllGiants extends React.Component {
+  render() {
+    return <Text>All Giants</Text>;
+  }
+}
